requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
playwright==1.40.0
requests==2.31.0
beautifulsoup4==4.12.2
pytesseract==0.3.10
pillow==10.1.0
python-multipart==0.0.6
pydantic==2.4.2 